fix: correct the DBus service so that it can see host schemas
The package was exporting a DBus service, which means that the `Exec=` line of the `.desktop` file is completely ignored.

This meant that every time you launch the application via GNOME, it's being launched via its DBus service instead of the desktop file's Exec line.

As a result, it was always launching without the wrapper script, and therefore couldn't see the host's schemas.

The service is now also patched, to use the proper launcher. Now both the `.desktop` file and the `.service` file point at the correct wrapper script for correct host integration in all scenarios.