fix: correctly prioritize the host's gschema instead of the flatpak's
The environment variable was being built in the wrong order, and was putting the Flatpak's data locations as the highest priority. This meant that dconf-editor showed *completely wrong* values for most settings in dconf, since it was looking at the default Flatpak environment rather than the host's dconf environment!

This has now been fixed to use the correct priority, putting all host paths first, and then the Flatpak's default locations after that as fallbacks.